% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{estimate.sigma}
\alias{estimate.sigma}
\title{estimates the sigma parameter of the kernel}
\usage{
estimate.sigma(sample)
}
\arguments{
\item{sample}{a sample matrix from the target density}
}
\description{
this function will try to estimate a scalar sigma, assuming that
the sample variables are all very similar to one another.
}
