% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{event.tf}
\alias{event.tf}
\title{A scheduled transformation event}
\usage{
event.tf(t, state.tf, param.tf)
}
\arguments{
\item{t}{the times at which transformations occur}

\item{state.tf}{the transformation that affects the state}

\item{param.tf}{the transformation that affects the parameters}
}
\value{
a list with two components: "time" and "tf";
tf itself is a list of two components: "state" and "param";
both are transformations with the components A and b.
}
\description{
This function creates a list of all necessary characteristics of a
repeating transformation event.
}
\details{
A scheduled event happens at a predetermined time t. It affects
both model state and parameters.

The transformations are affine (linear + shift). If no
transformation shall occur, an identity matrix can be used.
They are encoded as a series of matrices A and a series of
vectors b, both series have the same length as t. The third index
of A and b corresponds to the time: t\link{i}, A\link{,,i}, and b\link{,,i} belong
together.
}
