% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_gsl_odeiv2.R
\name{integrationMethod}
\alias{integrationMethod}
\title{Selects the integration method, by name}
\usage{
integrationMethod(
  charMethod = c("msbdf", "msadams", "bsimp", "rk4imp", "rk2imp", "rk1imp", "rk8pd",
    "rkck", "rkf45", "rk4", "rk2")
)
}
\arguments{
\item{charMethod}{a scalar character string, defaulting to msbdf}
}
\value{
an integer offset, which will select the named method in the solvers
}
\description{
These are the methods in the gsl library (documented in the
official documentation), but in reverse order, as they are
approximately ordered by complexity, with more complex methods
usually being better.
}
\details{
It is therefore a reasonable approach to try methods from the more
complex end of the list first and try the next method if the
solutions are too slow. But ensure checking the accuracy/stability of the
result. The mapping between method names and keys:

\if{html}{\out{<div class="sourceCode">}}\preformatted{     msbdf:	 0
   msadams:	 1
     bsimp:	 2
    rk4imp:	 3
    rk2imp:	 4
    rk1imp:	 5
     rk8pd:	 6
      rkck:	 7
     rkf45:	 8
       rk4:	 9
       rk2:	10
}\if{html}{\out{</div>}}

The returned value must be passed to the solver function calls, e.g.: r_gsl_odeiv2_outer(...,method = integrationMaethod("msbdf"))
}
