% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde.c}
\alias{kde.c}
\title{kernel density estimator}
\usage{
kde.c(sample, normalize = FALSE)
}
\arguments{
\item{sample}{a matrix with columns of observations of continuous
variables from a target density.}
}
\value{
A closure, function of x, where x is the matrix of columns
for which to estimate densities
}
\description{
written in C, the prupose of this function is to make very fast
kernel density estimates, with no user supplied bandwidth
parameters.
}
