% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{make.experiment}
\alias{make.experiment}
\title{Create a simulation experiment}
\usage{
make.experiment(t, y0, par, event.t = NULL, state.tf = NULL, param.tf = NULL)
}
\description{
Create a simulation experiment
}
