% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{rgsl.example}
\alias{rgsl.example}
\title{Locate Example Files}
\usage{
rgsl.example(modelName = "HarmonicOscillator", f = "_gvf.c")
}
\arguments{
\item{f}{file ending, search will be restricted to this}
}
\description{
This function returns the full path to the example model.
}
